requirements.txt
# RGB2Voxel 依赖库
# 安装命令: pip install -r requirements.txt

# 核心依赖
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
tqdm>=4.60.0
pyyaml>=6.0

# 3D处理
open3d>=0.15.0

# 深度学习
torch>=1.9.0
torchvision>=0.10.0

# SAM3相关
transformers>=4.30.0
pycocotools>=2.0.6

# 可选: 用于交互式可视化
# pynput>=1.7.0

